Update BrowserTracing import for tree shaking
In https://github.com/getsentry/sentry-javascript/pull/4204 we updated
the `BrowserTracing` import in the JS SDK to be exported individually.
We want users to use this individual import so that the other
integrations are treeshaken. This patch updates our documentation to
import `BrowserTracing` directly from `@sentry/tracing` instead of
through the `Integrations` object, which hopefully should lead to
a bundle size reduction for people.